fsk_distance_measure: initial version, lots of errors and warnings when compile

+++ embedded/fsk_distance_measure/app/distance_measure/inc/micromotion_detection.h
@@ -4,10 +4,10 @@
 #include "sys.h"
 
 typedef struct {
-	float PAD;
-	u16 cfar_N;
-	u16 pro_N;	
+    float PAD;
+    u16 cfar_N;
+    u16 pro_N;
 } cfar_param_t;
 
-int micromotion_detection(s16 *data, int len, float offset);
+int micromotion_detection(s16 *data, int len, s32 offset);
 #endif
